unify view models
